gbproxy: Set configuration in a more dynamic way
Let's generalize the data types a bit, and move the gb (bssgp) config
into a module parameter.  That parameter then is used for both the PCUs
as well as (concatenated) for the SGSN side.

This allows the configuration file to have more control over the number
of BVC within each NSE.

Change-Id: I43a3a8e133cf0f0e377b64d1b385e88285246957