BTS_Tests: test forwarding of SI1, SI3 and SI13 via PCUIF
Whenever the BSC is updating SI1, SI3 or SI13 via RSL, the PCU should be
informed about the change via PCUIF as well. For SI13 this is already
the case and a testcase exists. The version 11 of the PCUIF protocol is
now capable to update SI1 and SI3 as well.

- Update BTS_Tests.TC_pcu_ver_si13 so that it works with the current
  protocol version
- Add BTS_Tests.TC_pcu_ver_si3 and BTS_Tests.TC_pcu_ver_si1 that test
  SI1 and SI3 as well.

Depends: osmo-bts Ib7aeb41e634ad6fcab3766a4667b0267c749436a
Change-Id: I5138ab183793e7eee4dc494318d984e9f1f56932
Related: SYS#5103