fr: Add TC_ul2dl_ud
This test case tests NS-UL-UNITDATA transmit and expectes the exact same
LLC PDU to be sent back as NS-DL-UNITDATA.

This way we can verify that both uplink and downlink paths are working,
and that no messages are lost.  It works stable here on my laptop,
showing that if we test the TTCN3 NS + BSSGP code over a E1 line
against another instance (fr against fr-net) works reliable.

Change-Id: Ic115af02207c9b9f4c84fa75890048acb6856c79