gbproxy: Fix race condition checking for unblocked BVC
It sometimes happens that the ttcn3 test checks the BVC FSM state before
the gbproxy had time to process the BVC_RESET_ACK from the "SGSN". It
then reports the state WAIT_RESET_ACK which will fail the test.

Introduce a small wait to give the gbproxy some more time.

Change-Id: Ic31b2188195c5d76b2d5aac9fa1f0b4efa5ca8b2