PCU_Tests: verify deallocation of TBFs
In most tests, add a final stage that ensures that all TBFs are freed in
some way or other, by polling the talloc report via VTY.

This implies that each test run is using the osmo-pcu process
exclusively, i.e. that no other test cases run in parallel. In practice,
this is currently always the case.

Related: SYS#5209
Change-Id: Ied9deab5aff9b4b5f5874216393d6c56293754ab