bsc: make TC_stats_conn_fail less fragile
Add a delay between sending the RSL Conn Fail IND and checking the stats
values, so that osmo-bsc has a good chance of receiving the message and
promoting the failure into the stats values first.

Even though jenkins shows no failures, when testing locally, this test
fails a lot for me without that bit of delay.

Change-Id: Iaf0173239528337283b23f70840d36ad25f14950